+++ src/db/schema.ts
@@ -1,6 +1,5 @@
 import { pgTable, serial, text, integer, timestamp } from "drizzle-orm/pg-core";
 import type { InferSelectModel } from "drizzle-orm";
-import { number } from "zod";
 export const userTable = pgTable("user", {
   id: serial("id").primaryKey(),
   email: text("email").unique(),
